// app/requests/_personal/postUpdateOrderActivity/postUpdateOrderActivitySuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface PostUpdateOrderActivitySuccess {
  data: {
    id: number;
    selfEmployed: boolean;
    status: number;
    place: {
      id: number;
      name: string;
      latitude: string;
      longitude: string;
      address_kladr: string;
      logo: string;
      region: {
        id: number;
        name: string;
      };
      brand: {
        id: number;
        name: string;
        logo: null | string;
        description: string;
      };
    };
    user: {
      id: number;
      phone: number;
      email: string;
      logo: string;
      roles: {
        id: number;
        name:
          | "admin"
          | "client"
          | "manager"
          | "recruiter"
          | "specialist"
          | "supervisor";
      }[];
    };
    orderActivities: {
      id: number;
      viewActivity: {
        id: number;
        name: string;
        detailName: string;
        previewText: string;
        logo: string;
        traveling: boolean;
      };
      count: number;
      dateStart: string;
      dateEnd: string;
      needFoto: boolean;
      dateActivity: {
        timeStart: string;
        timeEnd: string;
        places: {
          id: number;
          name: string;
          latitude: string;
          longitude: string;
          address_kladr: string;
          logo?: null | string;
          region: {
            id: number;
            name: string;
          };
          brand?: null | {
            id: number;
            name: string;
            logo: null | string;
            description: null | string;
          };
        }[];
      }[];
    }[];
    acceptUser: null | {
      id: number;
      phone: number;
      email: string;
      logo: string;
    };
  };
}